boilerplate/python/signal-generator/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==4.22.3
redis==4.5.4
pika==1.3.1
psycopg2-binary==2.9.6
sqlalchemy==2.0.9
pandas==2.0.0
numpy==1.24.2
ta==0.10.2
pytest==7.3.1
httpx==0.24.0
tenacity==8.2.2
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-jaeger==1.17.0
prometheus-client==0.16.0
